% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciudades.R
\docType{data}
\name{ciudades}
\alias{ciudades}
\title{Dataset de ciudades}
\format{
Un tibble con N filas y X columnas:
\describe{
\item{columna1}{Descripción}
\item{columna2}{Descripción}
}
}
\source{
Archivo propio procesado desde \code{data-raw/ciudades.csv}
}
\usage{
ciudades
}
\description{
Datos de ciudades con informacion geografica y administrativa.
}
\keyword{datasets}
