% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrar_region.R
\name{filtrar_region}
\alias{filtrar_region}
\title{Filtrar ciudades por region}
\usage{
filtrar_region(region)
}
\arguments{
\item{region}{Cadena de texto indicando la region.}
}
\value{
Un tibble con las ciudades pertenecientes a la region indicada.
}
\description{
Esta funcion filtra el dataset interno \code{ciudades} segun la region
seleccionada. Las regiones validas son: "norte", "centro" y "sur".
}
\examples{
filtrar_region("norte")

# Region invalida (genera error)
\dontrun{
filtrar_region("patagonia")
}

}
