% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_ciudades.R
\name{leer_ciudades}
\alias{leer_ciudades}
\title{Leer ciudades desde los datos internos}
\usage{
leer_ciudades()
}
\value{
Un tibble con los datos del dataset \code{ciudades}.
}
\description{
Esta funcion devuelve el dataset interno \code{ciudades} como un tibble.
El dataset es cargado automaticamente desde la carpeta \code{data/}
cuando el paquete es cargado.
}
\examples{
# Devuelve el dataset de ciudades
leer_ciudades()

}
